Address PR review: consolidate VariantTests, fix AgileComPointer ownership in tests
Per the test conventions in .github/instructions/tests.instructions.md, fold VariantConversionTests, VariantToObjectTests, and VariantByrefTests into a single VariantTests so there is one FooTests per production type.

Switch AgileComPointer tests to 'using ComScope<IUnknown> source' + takeOwnership: false so the caller's ComScope is deterministically released and the GIT's AddRef from RegisterInterfaceInGlobal owns its own ref.
